backend/requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
psycopg2==2.9.3
gunicorn==20.1.0
